+++ PurchaseSaleRecord.cpp
@@ -41,8 +41,8 @@ void PurchaseSaleRecord::load(std::ifstream &fin) {
 }
 
 // Computes the profit/loss for the PurchaseSaleRecord
-int PurchaseSaleRecord::computeProfitLoss() {
-    int totInvestment = (_numShares * _purchasePrice) + _purchaseFee + _saleFee;
+double PurchaseSaleRecord::computeProfitLoss() {
+    double totInvestment = (_numShares * _purchasePrice) + _purchaseFee + _saleFee;
 
     return (_numShares * _salePrice) - totInvestment;
 }
